implement remote resolver
